UAS
Final

+++ application/controllers/admin/Berita.php
@@ -20,19 +20,33 @@ class Berita extends CI_Controller {
         $data["kategori"] = $this->kategori_model->show_kategori();
         $data['user'] = $this->user_model->show_user();
         $data["konten"] = "page/admin/berita/index";
-        $this->load->view("layouts/main2", $data);
+        $this->load->view("layouts/admin", $data);
     }
 
     public function add() {
         $data["kategori"] = $this->kategori_model->show_kategori();
-        if(isset($_POST) && count($_POST) > 0){   
+        if(isset($_POST) && count($_POST) > 0){
+            $config['upload_path']      = './assets/img/thumbnail/';
+            $config['allowed_types']    = '*';
+            $config['max_size']         = 10240;
+            $config['encrypt_name']     = true;
+
+            $this->load->library('upload', $config);
+            if ($this->upload->do_upload('foto')){
+                $img = $this->upload->data();
+            }
+            else{
+                $img = '';
+            }
+            //var_dump($_POST);
             $store = array(
 				'brt_admin' => $this->input->post('brt_admin'),
 				'brt_create' => $this->input->post('brt_tgl'),
 				'brt_tgl' => $this->input->post('brt_tgl'),
 				'brt_judul' => $this->input->post('brt_judul'),
 				'brt_isi' => $this->input->post('brt_isi'),
-				'brt_kategori' => $this->input->post('brt_kategori'),
+				'brt_kategori' => implode(",",$this->input->post('brt_kategori')),
+				'brt_foto' => $img['file_name'],
             );
             try{
                 $berita_id = $this->berita_model->add_berita($store);
@@ -47,7 +61,7 @@ class Berita extends CI_Controller {
             $data["mBerita"] = true;
             $data["judul"] = "Tambah berita";
             $data["konten"] = "page/admin/berita/add";
-            $this->load->view("layouts/main2", $data);
+            $this->load->view("layouts/admin", $data);
         }
     }
 
@@ -55,13 +69,27 @@ class Berita extends CI_Controller {
         $data['berita'] = $this->berita_model->get_berita($id);
         $data["kategori"] = $this->kategori_model->show_kategori();
         if(isset($data['berita']['brt_id'])){
-            if(isset($_POST) && count($_POST) > 0) {   
+            if(isset($_POST) && count($_POST) > 0) {
+                $config['upload_path']      = './assets/img/thumbnail/';
+                $config['allowed_types']    = '*';
+                $config['max_size']         = 10240;
+                $config['encrypt_name']     = true;
+    
+                $this->load->library('upload', $config);
+                if ($this->upload->do_upload('foto')){
+                    $img = $this->upload->data();
+                }
+                else{
+                    $img['file_name'] =  $this->input->post('old');
+                }
                 $update = array(
                     'brt_admin' => $this->input->post('brt_admin'),
                     'brt_create' => $this->input->post('brt_create'),
+                    'brt_tgl' => $this->input->post('brt_tgl'),
                     'brt_judul' => $this->input->post('brt_judul'),
                     'brt_isi' => $this->input->post('brt_isi'),
-                    'brt_kategori' => $this->input->post('brt_kategori'),
+                    'brt_kategori' => implode(",",$this->input->post('brt_kategori')),
+                    'brt_foto' => $img['file_name'],
                 );
                 
                 try{
@@ -77,7 +105,7 @@ class Berita extends CI_Controller {
                 $data["mBerita"] = true;
                 $data["judul"] = "Edit berita";
                 $data["konten"] = "page/admin/berita/edit";
-                $this->load->view("layouts/main2", $data);
+                $this->load->view("layouts/admin", $data);
             }
         }
         else {
@@ -90,12 +118,17 @@ class Berita extends CI_Controller {
         $berita = $this->berita_model->get_berita($id);
         
         if(isset($berita['brt_id'])){
-            try{
-                $this->berita_model->delete_berita($id);
-                $this->session->set_flashdata('success', 'Anda berhasil menhapus data');
+            if($this->session->id_u==$berita['brt_admin']||$this->session->status==1){
+                try{
+                    $this->berita_model->delete_berita($id);
+                    $this->session->set_flashdata('success', 'Anda berhasil menhapus data');
+                }
+                catch(Exception $eror){
+                    $this->session->set_flashdata('fail', 'Anda gagal menghapus data');
+                }
             }
-            catch(Exception $eror){
-                $this->session->set_flashdata('fail', 'Anda gagal menghapus data');
+            else{
+                $this->session->set_flashdata('fail', 'Hanya Admin atau sang penulis yang dapat menhapus berita ini');
             }
             redirect('admin/berita/index');
         }
@@ -113,6 +146,6 @@ class Berita extends CI_Controller {
         $data['user'] = $this->user_model->show_user();
         $data["konten"] = "page/admin/berita/show";
         // $data["konten"] = "page/InBuild";
-        $this->load->view("layouts/main2", $data);
+        $this->load->view("layouts/admin", $data);
     }
 }
